refactor: extract and unify partition and goal event selector makers
