+++ src/chart_types/goal_chart/state/selectors/on_element_over_caller.ts
@@ -21,38 +21,13 @@ import createCachedSelector from 're-reselect';
 import { Selector } from 'react-redux';
 
 import { ChartTypes } from '../../..';
-import { SeriesIdentifier } from '../../../../common/series_id';
+import { getOnElementOverSelector } from '../../../../common/event_handler_selectors';
 import { LayerValue } from '../../../../specs';
 import { GlobalChartState } from '../../../../state/chart_state';
 import { getChartIdSelector } from '../../../../state/selectors/get_chart_id';
 import { getSettingsSpecSelector } from '../../../../state/selectors/get_settings_specs';
 import { getSpecOrNull } from './goal_spec';
 import { getPickedShapesLayerValues } from './picked_shapes';
-
-function isOverElement(prevPickedShapes: Array<Array<LayerValue>> = [], nextPickedShapes: Array<Array<LayerValue>>) {
-  if (nextPickedShapes.length === 0) {
-    return;
-  }
-  if (nextPickedShapes.length !== prevPickedShapes.length) {
-    return true;
-  }
-  return !nextPickedShapes.every((nextPickedShapeValues, index) => {
-    const prevPickedShapeValues = prevPickedShapes[index];
-    if (prevPickedShapeValues === null) {
-      return false;
-    }
-    if (prevPickedShapeValues.length !== nextPickedShapeValues.length) {
-      return false;
-    }
-    return nextPickedShapeValues.every((layerValue, i) => {
-      const prevPickedValue = prevPickedShapeValues[i];
-      if (!prevPickedValue) {
-        return false;
-      }
-      return layerValue.value === prevPickedValue.value && layerValue.groupByRollup === prevPickedValue.groupByRollup;
-    });
-  });
-}
 
 /**
  * Will call the onElementOver listener every time the following preconditions are met:
@@ -61,32 +36,13 @@ function isOverElement(prevPickedShapes: Array<Array<LayerValue>> = [], nextPick
  * @internal
  */
 export function createOnElementOverCaller(): (state: GlobalChartState) => void {
-  let prevPickedShapes: Array<Array<LayerValue>> = [];
+  const prev: { pickedShapes: LayerValue[][] } = { pickedShapes: [] };
   let selector: Selector<GlobalChartState, void> | null = null;
   return (state: GlobalChartState) => {
     if (selector === null && state.chartType === ChartTypes.Goal) {
       selector = createCachedSelector(
         [getSpecOrNull, getPickedShapesLayerValues, getSettingsSpecSelector],
-        (spec, nextPickedShapes, settings): void => {
-          if (!spec) {
-            return;
-          }
-          if (!settings.onElementOver) {
-            return;
-          }
-
-          if (isOverElement(prevPickedShapes, nextPickedShapes)) {
-            const elements = nextPickedShapes.map<[Array<LayerValue>, SeriesIdentifier]>((values) => [
-              values,
-              {
-                specId: spec.id,
-                key: `spec{${spec.id}}`,
-              },
-            ]);
-            settings.onElementOver(elements);
-          }
-          prevPickedShapes = nextPickedShapes;
-        },
+        getOnElementOverSelector(prev),
       )({
         keySelector: getChartIdSelector,
       });
